Feature ETP-2308: Add Jenkins pipeline for Tomcat service checks
This commit introduces a new Jenkinsfile that defines a pipeline for
building and deploying the Tomcat service. It includes stages for
waiting for Docker-in-Docker (DinD) readiness, building the environment,
and checking the Tomcat status. The pipeline ensures that the necessary
environment variables and configurations are set up for a successful
deployment.